requirements.txt
pip~=21.3.1
attrs~=21.4.0
wheel~=0.37.1
Pillow~=9.2.0
Jinja2~=3.1.1
setuptools~=60.2.0
Django~=4.0.3
pytz~=2022.1
asgiref~=3.5.0
sqlparse~=0.4.2
mysqlclient~=2.1.1
colorama~=0.4.4
PyYAML~=6.0
numpy~=1.23.3
requests~=2.27.1
packaging~=21.3
pyparsing~=3.0.9
zipp~=3.8.0
djangorestframework~=3.13.1